refactor(server): expose runtime app assembly
